Added full 32-bit target compatibility.
size_t and ssize_t are only 32 bit on 32 bit architectures (at least in Mac OS X), so they have to be used carefully.

The basic gist of what I did was this:

Anything having to do with volume/file sizes and/or offsets in **bytes**, use 64-bit variants: a volume's size in bytes is an `uint64_t`, offsets in file reads are an `off_t` (int64_t). Also, when necessary, be sure to cast to 64-bit variants during multiplication to avoid integer overflow.  File/volume reads/writes still use size_t to specify size of the buffer, which is appropriate.

Anything having to do with volume/file sizes in **blocks**, use `size_t`/`ssize_t`, since HFS+ itself uses 32 bits for a volume's total block count.

+++ src/cdefs.h
@@ -56,7 +56,7 @@ typedef char uuid_string_t[37];
 
 #else                           // !GC_ENABLED
 
-#define ALLOC(buf_size)        calloc(1, (buf_size))
+#define ALLOC(buf_size)        calloc(1, (size_t)(buf_size))
 #define REALLOC(buf, buf_size) realloc((buf), (buf_size))
 #define FREE(buf)              free((buf))
